+++ test/extensions/filters/network/redis_proxy/command_lookup_speed_test.cc
@@ -29,6 +29,8 @@ public:
   NoOpSplitCallbacks() {}
   ~NoOpSplitCallbacks() {}
 
+  bool connectionAllowed() override { return true; }
+  void onAuth(const std::string&) override {}
   void onResponse(Common::Redis::RespValuePtr&&) override {}
 };